chore: separate e2e workflow
