+++ fastcrypto/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fastcrypto"
-version = "0.1.6"
+version = "0.1.7"
 license = "Apache-2.0"
 authors = ["Mysten Labs <build@mystenlabs.com>"]
 edition = "2021"
